42-freeze.patchtches.xen/30-bit-field-booleans.patch: Packet loss in DomU -xen netback driver (306896).
suse-commit: fbd1c55158ff4be5a8d394d64e4546681409bd6b

+++ drivers/xen/netback/common.h
@@ -78,8 +78,8 @@ typedef struct netif_st {
 	int features;
 
 	/* Internal feature information. */
-	int can_queue:1;	/* can queue packets for receiver? */
-	int copying_receiver:1;	/* copy packets to receiver?       */
+	u8 can_queue:1;	/* can queue packets for receiver? */
+	u8 copying_receiver:1;	/* copy packets to receiver?       */
 
 	/* Allow netif_be_start_xmit() to peek ahead in the rx request ring. */
 	RING_IDX rx_req_cons_peek;
